requirements.txt
confluent-kafka==1.8.2
airflow-provider-kafka==0.2.1
astronomer-providers[all]==1.12.0
apache-airflow-providers-slack==7.2.0